refactor(inventory): move delete state into useItemsPageModel
Moves deletingItemId + deleteMutation into the model hook to keep
ItemsPage under the max-lines-per-function lint limit introduced when
the persistent Add Item button was added. Model now owns all tRPC and
local state; component only renders.
