+++ packages/app-inventory/src/pages/ItemsPage.tsx
@@ -1,8 +1,6 @@
 import { Plus } from 'lucide-react';
-import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { trpc } from '@pops/api-client';
 import {
   AlertDialog,
   AlertDialogAction,
@@ -63,15 +61,6 @@ export function ItemsPage() {
   const { filters, navigate } = model;
   const hasSearchOrFilters = !!filters.search || model.hasActiveFilters;
 
-  const [deletingItemId, setDeletingItemId] = useState<string | null>(null);
-  const utils = trpc.useUtils();
-  const deleteMutation = trpc.inventory.items.delete.useMutation({
-    onSuccess: () => {
-      void utils.inventory.items.list.invalidate();
-      setDeletingItemId(null);
-    },
-  });
-
   const addButton = (
     <Button onClick={() => navigate('/inventory/items/new')} prefix={<Plus className="h-4 w-4" />}>
       Add Item
@@ -112,13 +101,15 @@ export function ItemsPage() {
         locationPathMap={model.locationPathMap}
         onOpen={(id) => navigate(`/inventory/items/${id}`)}
         onEdit={(id) => navigate(`/inventory/items/${id}/edit`)}
-        onDeleteRequest={setDeletingItemId}
+        onDeleteRequest={model.setDeletingItemId}
       />
       <DeleteItemDialog
-        isOpen={deletingItemId !== null}
-        isPending={deleteMutation.isPending}
-        onConfirm={() => deletingItemId && deleteMutation.mutate({ id: deletingItemId })}
-        onClose={() => setDeletingItemId(null)}
+        isOpen={model.deletingItemId !== null}
+        isPending={model.deleteMutation.isPending}
+        onConfirm={() =>
+          model.deletingItemId && model.deleteMutation.mutate({ id: model.deletingItemId })
+        }
+        onClose={() => model.setDeletingItemId(null)}
       />
     </div>
   );
